requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
python-docx
pandas
tqdm
transformers
torch
